+++ src/main/java/com/danit/erp/repository/GroupScheduleRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupScheduleRepository extends JpaRepository<GroupSchedule, Long> {
+public interface GroupScheduleRepository extends JpaRepository<GroupSchedule, Integer> {
 
   Optional<GroupSchedule> findByDayOfWeek(String group);
 }
