+++ src/main/java/com/danit/erp/repository/dictionary/EducationRepository.java
@@ -9,11 +9,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducationRepository extends JpaRepository<Education, Long> {
+public interface EducationRepository extends JpaRepository<Education, Short> {
   Optional<Education>findBySpecialization(String name);
   List<Education> findByDeletedFalse();
   Page<Education> findByDeletedFalse(Pageable pageable);
-  Optional<Education> findByIdAndDeletedFalse(Long userId);
+  Optional<Education> findByIdAndDeletedFalse(Short userId);
 
 
 }
