+++ src/main/java/com/danit/erp/repository/dictionary/GroupRepository.java
@@ -9,14 +9,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
   Optional<Group> findByGroupName(String nameGroup);
 
   List<Group> findByDeletedFalse();
 
   Page<Group> findByDeletedFalse(Pageable pageable);
 
-  Optional<Group> findByIdAndDeletedFalse(Long userId);
+  Optional<Group> findByIdAndDeletedFalse(Integer userId);
 
 
 }
