+++ src/main/java/com/danit/erp/repository/dictionary/ProfessionRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfessionRepository extends JpaRepository<Profession, Long> {
+public interface ProfessionRepository extends JpaRepository<Profession, Short> {
   List<Profession> findByDeletedFalse();
   Page<Profession> findByDeletedFalse(Pageable pageable);
-  Optional<Profession> findByIdAndDeletedFalse(Long userId);
+  Optional<Profession> findByIdAndDeletedFalse(Short userId);
 
   Optional<Profession>  findByName(String name);
 }
