+++ src/main/java/com/danit/erp/repository/dictionary/UniversityRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversityRepository extends JpaRepository<University, Long> {
+public interface UniversityRepository extends JpaRepository<University, Integer> {
   List<University> findByDeletedFalse();
 
-  Optional<University> findByIdAndDeletedFalse(Long userId);
+  Optional<University> findByIdAndDeletedFalse(Integer userId);
 
   Page<University> findByDeletedFalse(Pageable pageable);
