+++ src/main/java/com/danit/erp/repository/dictionary/roles/CoordinatorRepository.java
@@ -9,12 +9,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoordinatorRepository extends JpaRepository<Coordinator, Long> {
+public interface CoordinatorRepository extends JpaRepository<Coordinator, Integer> {
   Optional<Coordinator> findByFullName(String fullName);
   Page<Coordinator> findByDeletedFalse(Pageable pageable);
 
   List<Coordinator> findByDeletedFalse();
-  Optional<Coordinator> findByIdAndDeletedFalse(Long userId);
+  Optional<Coordinator> findByIdAndDeletedFalse(Integer userId);
 
 
 }
