+++ src/main/java/com/danit/erp/repository/dictionary/roles/RoleRepository.java
@@ -9,13 +9,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Short> {
   Optional<Role> findByRole(String role);
   Page<Role> findByDeletedFalse(Pageable pageable);
 
   List<Role> findByDeletedFalse();
 
-  Optional<Role> findByIdAndDeletedFalse(Long userId);
+  Optional<Role> findByIdAndDeletedFalse(Short userId);
 
 
 }
